+++ src/components/board/index.js
@@ -1,0 +1,3 @@
+import Board from './Board';
+
+export default Board;
